cleaning up mds-types
